Upgrade to verision 0.0.0.12, add FPU code into corresponding folders

+++ Drivers/ADC102S021.cs
@@ -45,7 +45,7 @@ namespace Drivers
 
                     if (i > 3)
                     {
-                        current = (ushort)(current << 1 | DOUT);
+                        current = (ushort)(current << 1 | (byte)DOUT);
                     }
 
                     internalSCLK = true;
